Refactored code and cfixed bugs
