+++ webapp/src/components/SurveyResultsNov30.jsx
@@ -3,25 +3,25 @@ import React from 'react';
 const SurveyResults = ({ submittedData, noisyData }) => {
   const renderSurveyResult = (label, submittedValue, noisyValue) => (
     <div key={label}>
-      <strong>{label}:</strong> {submittedValue || ''} (Noisy: {noisyValue || ''})
+      <strong>{label}:</strong> {submittedValue || ''}  || {noisyValue || ''}
     </div>
   );
 
   return (
     <>
       <div className="panel small-panel">
-        <h3>Submitted Data</h3>
+        <h3>Submitted || Noisy</h3>
         {Object.entries(submittedData).map(([label, value]) =>
           renderSurveyResult(label, value, noisyData[label])
         )}
       </div>
 
-      <div className="panel small-panel">
+      { /* <div className="panel small-panel">
         <h3>Noisy Data</h3>
         {Object.entries(noisyData).map(([label, value]) =>
           renderSurveyResult(label, submittedData[label], value)
         )}
-      </div>
+      </div> */ }
     </>
   );
 };
