+++ webapp/src/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import 'bootstrap/dist/css/bootstrap.min.css';
 //import App from './App';
-import App from './Nov30/AppNov30';
+import App from './AppNov30';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
